requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap4==2024.10
Django==5.2.1
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-rosetta==0.10.2
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
model-bakery==1.20.4
polib==1.2.0
psycopg2-binary==2.9.10
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
